requirements.txt
numpy==1.26.4
Pillow==10.2.0
scipy==1.12.0
